Update on "Speed up calculate Qparams for per-channel observers"
Use vectorization to speed up calculate Qparams for per-channel observers. New implementation is about 1000 times faster.

Test plan:
```
import torch
import time
import numpy as np
from torch.quantization.observer import PerChannelMinMaxObserver

obs = PerChannelMinMaxObserver()
acc_time = 0
X = torch.randn(1000, 10)
obs(X)
for i in range(100):
    start = time.time()
    obs.calculate_qparams()
    acc_time = acc_time + time.time()-start
print(acc_time)
```
Before change:
20.3
After change:
0.017

Task:
https://github.com/pytorch/pytorch/issues/30348#event-2824868602

Differential Revision: [D18711905](https://our.internmc.facebook.com/intern/diff/D18711905/)

[ghstack-poisoned]

+++ torch/quantization/observer.py
@@ -126,7 +126,7 @@ class _ObserverBase(ObserverBase):
 
     @torch.jit.export
     def _calculate_qparams(self, min_val, max_val):
-        # type: (Optional[Tensor], Optional[Tensor]) -> Tuple[Tensor, Tensor, int]
+        # type: (Optional[Tensor], Optional[Tensor]) -> Tuple[Tensor, Tensor]
         r"""Calculates the per channel quantization parameters, given min and max
         value tensors.
 
@@ -174,12 +174,12 @@ class _ObserverBase(ObserverBase):
         if self.qscheme == torch.per_tensor_symmetric or self.qscheme == torch.per_channel_symmetric:
             max_val = torch.max(-min_val, max_val)
             scale = max_val / (float(qmax - qmin) / 2)
-            scale = torch.max(scale, torch.tensor(1.0).new_full(scale.size(), self.eps))
+            scale = torch.max(scale, torch.tensor(1.0).new_full(scale.size(), self.eps, dtype=torch.float32))
             if self.dtype == torch.quint8:
                 zero_point = zero_point.new_full(zero_point.size(), 128)
         else:
             scale = (max_val - min_val) / float(qmax - qmin)
-            scale = torch.max(scale, torch.tensor(1.0).new_full(scale.size(), self.eps))
+            scale = torch.max(scale, torch.tensor(1.0).new_full(scale.size(), self.eps, dtype=torch.float32))
             zero_point = qmin - torch.round(min_val / scale)
             zero_point = torch.max(zero_point, torch.tensor(1.0).new_full(zero_point.size(), qmin))
             zero_point = torch.min(zero_point, torch.tensor(1.0).new_full(zero_point.size(), qmax))
